requirements.txt
appdirs==1.4.4
certifi==2016.2.28
cffi==1.10.0
cycler==0.10.0
Cython==0.29.21
dataclasses==0.8
decorator==4.4.2
distro==1.5.0
easydict==1.9
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.3
mkl-fft==1.0.6
mkl-random==1.0.1
netron==4.8.2
numpy==1.19.5
onnx==1.4.0
opencv-python==4.5.1.48
packaging==20.8
Pillow==8.1.0
protobuf==3.15.6
pycocotools==2.0.0
pycparser==2.18
pycuda==2020.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytools==2021.2
PyYAML==5.4.1
scikit-build==0.11.1
six==1.15.0
TBB==0.1
torch==1.2.0
torchvision==0.2.0
tqdm==4.61.0
typing==3.7.4.3
typing-extensions==3.7.4.3
